requirements.txt
pandas
cx_Oracle
pymysql
sqlalchemy
sentry-sdk